% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{NGScopyData-package}
\alias{NGScopyData}
\alias{NGScopyData-package}
\title{Subset of BAM files of human tumor and pooled normal sequencing data (Zhao et al. 2014) for the NGScopy package}
\usage{
tps_90.chr6()
tps_27.chr6()
tps_N8.chr6()
}
\description{
Subset of BAM files of human tumor and pooled normal samples by targeted panel sequencing (Zhao et al. 2014). Each sample is a 10 percent random subsample drawn from the original sequencing data. The pooled normal sample has been rescaled accroding to the total number of normal samples in the "pool". Here provided is the subsampled data on chr6 (hg19).
}
\author{
Xiaobei Zhao
}
\references{
Zhao et al (2014), Targeted Sequencing in Non-Small Cell Lung Cancer (NSCLC) Using the University of North Carolina (UNC) Sequencing Assay Captures Most Previously Described Genetic Aberrations in NSCLC. \emph{In preparation}
}
\seealso{
NGScopy
}
\keyword{package}

