% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tps_27.chr6}
\alias{tps_27.chr6}
\title{A subset of tumor sample (ID: 27) by targeted panel sequencing}
\usage{
tps_27.chr6()
}
\value{
character, the path of the (sorted) bam file and its index file
}
\description{
A subset of tumor sample (ID: 27) by targeted panel sequencing, a 10 percent random subsample drawn from chr6, hg19 (Zhao et al. 2014).
}
\examples{
require(NGScopyData)
tps_27.chr6()
}
\author{
Xiaobei Zhao
}
\seealso{
\link{NGScopyData} NGScopy
}

